==> Final-Solution/Notebooks/1-explore.ipynb <==
import pandas as pd 
import numpy as np 

path = "../Datasets/RAW/DxA23' Veri Seti.xlsx"

maaliyetler = pd.read_excel(path, sheet_name="İlçe Bazlı Maliyetler", skipfooter=0)
nufus = pd.read_excel(path, sheet_name="Mahalle Bazlı Nüfus Bilgisi", skipfooter=0)
brewstation = pd.read_excel(path, sheet_name="BrewStation Lokasyonları", skipfooter=0)

maaliyetler.head()

nufus.head()

#set upper case
maaliyetler = maaliyetler.apply(lambda x: x.astype(str).str.upper())
nufus = nufus.apply(lambda x: x.astype(str).str.upper())
brewstation = brewstation.apply(lambda x: x.astype(str).str.upper())

maaliyetler = maaliyetler[['İLÇE', 'KAHVE DÜKKANI AÇMA MALİYETİ']]
maaliyetler.columns = ['ilce', 'maliyet']

nufus = nufus[['İLÇE', 'MAHALLE', 'NÜFUS YOĞUNLUĞU (KİŞİ/KM2)', 'NÜFUS', 'YÜZÖLÇÜMÜ(KM2)']]
nufus.columns = ['ilce', 'mahalle', 'nufus_yogunlugu', 'nufus', 'yuzolcumu']

brewstation = brewstation[['ŞEHİR', 'AÇIK ADRES', 'İLÇE', 'MAHALLE']]
brewstation.columns = ['sehir', 'acik_adres', 'ilce', 'mahalle']

<hr>

turkish_chars = {
    'ü': 'u',
    'Ü': 'U',
    'ğ': 'g',
    'Ğ': 'G',
    'ş': 's',
    'Ş': 'S',
    'ı': 'i',
    'İ': 'I',
    'ö': 'o',
    'Ö': 'O',
    'ç': 'c',
    'Ç': 'C'
}

nufus['ilce'] = nufus['ilce'].replace(turkish_chars, regex=True)
nufus['mahalle'] = nufus['mahalle'].replace(turkish_chars, regex=True)

nufus.head()

maaliyetler['ilce'] = maaliyetler['ilce'].replace(turkish_chars, regex=True)
maaliyetler.head()

brewstation['ilce'] = brewstation['ilce'].replace(turkish_chars, regex=True)
brewstation['mahalle'] = brewstation['mahalle'].replace(turkish_chars, regex=True)
brewstation.head()

df = pd.merge(nufus, maaliyetler, on='ilce', how='left')
df.head()

df.isnull().sum()

df[df['maliyet'].isnull()]['ilce'].unique()

nufus['ilce'] = nufus['ilce'].replace('EYUP', 'EYUPSULTAN')

df = pd.merge(nufus, maaliyetler, on='ilce', how='left')
df.isnull().sum().sum()

df.to_csv('../Datasets/PROCESSED/nufus_maaliyet.csv', index=False)

df2 = pd.merge(brewstation, df, on=['ilce', 'mahalle'], how='left')

df2.isnull().sum()

df2[df2['nufus_yogunlugu'].isnull()]['ilce'].unique()

df2[df2['nufus_yogunlugu'].isnull()]['mahalle'].unique()

df2[df2['nufus_yogunlugu'].isnull()][['ilce', 'mahalle']].drop_duplicates()

brewstation['ilce'] = brewstation['ilce'].replace('EYUP', 'EYUPSULTAN')
brewstation['ilce'] = brewstation['ilce'].replace('EYUP ', 'EYUPSULTAN')
brewstation['mahalle'] = brewstation['mahalle'].replace('BARBAROS ', 'BARBAROS')

df[(df['ilce'].isin(['ATASEHIR', 'EYUP', 'BEYOGLU', 'EYUPSULTAN', 'TUZLA'])&
    (df['mahalle'].isin(['BARBAROS', 'TOPCULAR', 'KEMENKES KARAMUSTAFAPASA',
       'KEMENKES KARA MUSTAFA PASA', 'GOKTURK MERKEZ', 'DEFTERDAR',
       'EMNIYETTEPE', 'FATIH ', 'YESILPINAR'])))]

df2 = pd.merge(brewstation, df, on=['ilce', 'mahalle'], how='left')

df2.isnull().sum()

df2[df2['nufus_yogunlugu'].isnull()][['ilce', 'mahalle']].drop_duplicates()

df2[df2['nufus_yogunlugu'].isnull()]['ilce'].unique()

df2[df2['nufus_yogunlugu'].isnull()]['mahalle'].unique()

brewstation['mahalle'] = brewstation['mahalle'].replace('FATIH ', 'FATIH')
brewstation['mahalle'] = brewstation['mahalle'].replace('KEMENKES KARAMUSTAFAPASA', 'KEMANKES KARA MUSTAFA PASA')
brewstation['mahalle'] = brewstation['mahalle'].replace('KEMENKES KARA MUSTAFA PASA', 'KEMANKES KARA MUSTAFA PASA')

df[df['ilce'].isin(['TUZLA', 'BEYOGLU'])&
   df['mahalle'].isin(["KEMENKES KARA MUSTAFA PASA", 'FATIH'])]

#find mahalles start with 'K'
df[df['mahalle'].str.startswith('K')]['mahalle'].unique()

df2 = pd.merge(brewstation, df, on=['ilce', 'mahalle'], how='left')
df2.isnull().sum()

df2.to_csv('../Datasets/PROCESSED/brewstation.csv', index=False)
