==> Final-Solution/Notebooks/2-web-data-preprocessing.ipynb <==
import pandas as pd 
import numpy as np

ses = pd.read_excel('../Datasets/EXTERNAL/2016-yl-mahallem-istanbul-ses-sosyo-ekonomik-statu-skorlar.xlsx')
ses.head()

ses.to_csv('../Datasets/PROCESSED/ses_skorlari.csv', index=False)

<hr>

import chardet

with open('../Datasets/EXTERNAL/2017-yl-mahalle-bazl-bina-saylar.csv', encoding='iso-8859-9') as f:
    lines = f.readlines()

# split on ';' and strip whitespace from ends of lines
data = [line.strip().split(';') for line in lines]

header = data[0]
data = data[1:]

result = []

for row in data:
    dict_row = {}
    for i in range(len(header)):
        dict_row[header[i]] = row[i]
    result.append(dict_row)

konut_df = pd.DataFrame()
for dict in result:
    konut_df = pd.concat([konut_df, pd.DataFrame(dict, index=[0])])

konut_df.head()

konut_df.to_csv('../Datasets/PROCESSED/konut_sayilari.csv', index=False)

<hr>

with open('../Datasets/EXTERNAL/duraklar.csv', encoding='iso-8859-9') as f:
    lines = f.readlines()
    
data = [line.strip().split(',') for line in lines]

header = data[0]
data = data[1:]

result = []

for row in data:
    dict_row = {}
    for i in range(len(header)):
        dict_row[header[i]] = row[i]
    result.append(dict_row)

duraklar_df = pd.DataFrame()
for dict in result:
    duraklar_df = pd.concat([duraklar_df, pd.DataFrame([dict])])

duraklar_df.head()

duraklar_df = duraklar_df[['stop_id', 'stop_code', 'stop_lat', 'stop_lon']]
duraklar_df.to_csv('../Datasets/PROCESSED/duraklar.csv', index=False)

<hr>

otopark = pd.read_csv('../Datasets/EXTERNAL/ispark_parking.csv')
otopark.head()

otopark['COUNTY_NAME'].unique()

otopark = otopark[['PARK_TYPE_DESC', 'CAPACITY_OF_PARK', 'WORKING_TIME','COUNTY_NAME','LONGITUDE','LATITUDE']]
otopark.columns = ['park_tipi', 'kapasite', 'calisma_saati', 'ilce', 'longitude', 'latitude']
otopark.head(3)

otopark[otopark['calisma_saati']=='24 Saat'] = '00:00-23:59'
otopark['calisma_saati'].unique()

otopark[['acilis_saati', 'kapanis_saati']] = otopark['calisma_saati'].str.split('-', expand=True)
otopark.loc[otopark['kapanis_saati']=='00:00', 'kapanis_saati'] = '23:59'
otopark.head(3)

import datetime as dt

otopark['acilis_saati'] = otopark['acilis_saati'].apply(lambda x: dt.datetime.strptime(x.strip(), '%H:%M').time())
otopark['kapanis_saati'] = otopark['kapanis_saati'].apply(lambda x: dt.datetime.strptime(x.strip(), '%H:%M').time())
otopark.head(3)

calisma_suresi = []
for row in range(len(otopark)):
    acilis = otopark['acilis_saati'][row].hour
    kapanis = otopark['kapanis_saati'][row].hour
        
    if acilis > kapanis:
        calisma_suresi.append(acilis - kapanis + 12)
    
    else:
        calisma_suresi.append(kapanis - acilis)
otopark['calisma_suresi'] = calisma_suresi
otopark.head(3)

calisma_tipi = []

for row in range(len(otopark)):
    if otopark.iloc[row]['calisma_suresi'] >= 17:
        calisma_tipi.append('tum_gun')
    elif dt.time(6,0,0) <= otopark.iloc[row]['acilis_saati'] <= dt.time(19,0,0):
        calisma_tipi.append('gunduz')
    elif dt.time(18,0,0) < otopark.iloc[row]['acilis_saati'] or otopark.iloc[row]['acilis_saati'] < dt.time(6,0,0):
        calisma_tipi.append('gece')
    else:
        calisma_tipi.append('bilinmiyor')

    
otopark['calisma_tipi'] = calisma_tipi

otopark.head(3)

otopark['calisma_tipi'].value_counts()

otopark = otopark.drop(['calisma_saati','ilce'], axis=1)
otopark.to_csv('../Datasets/PROCESSED/otopark.csv', index=False)

<hr>

s_yardım = pd.read_excel('../Datasets/EXTERNAL/mahallelere-gore-sosyal-yardm-alan-hane-says.xlsx')
s_yardım.head()

turkish_chars = {
    'ü': 'u',
    'Ü': 'U',
    'ğ': 'g',
    'Ğ': 'G',
    'ş': 's',
    'Ş': 'S',
    'ı': 'i',
    'İ': 'I',
    'ö': 'o',
    'Ö': 'O',
    'ç': 'c',
    'Ç': 'C'
}

s_yardım['İLÇE'] = s_yardım['İLÇE'].replace(turkish_chars, regex=True)
s_yardım['MAHALLE'] = s_yardım['MAHALLE'].replace(turkish_chars, regex=True)
s_yardım.head(3)

s_yardım.to_csv('../Datasets/PROCESSED/sosyal_yardim.csv', index=False)

<hr>

with open('../Datasets/EXTERNAL/raylı-sistem-yolcu-segmenti.csv', encoding='iso-8859-9') as f:
    lines = f.readlines()
    
data = [line.strip().split(',') for line in lines]

columns = ["passage_cnt","passanger_cnt","transaction_year",
           "transaction_month","transaction_day","line",
           "station_name","station_number","town",
           "age","longitude","latitude"]

metro = pd.DataFrame(data[1:], columns=columns)
metro.head(2)

metro.drop(['line'], axis=1, inplace=True)

metro.head(12)

metro['transaction_year'] = metro['transaction_year'].astype(int)
metro['transaction_month'] = metro['transaction_month'].astype(int)
metro['transaction_day'] = metro['transaction_day'].astype(int)


metro['transaction_datetime'] = metro.apply(
    lambda x: dt.datetime(x['transaction_year'], x['transaction_month'], x['transaction_day']), axis=1)

metro.head(3)

metro = metro[['passage_cnt','passanger_cnt','age','longitude','latitude','transaction_datetime']]
metro.columns = ['gecis_sayisi','yolcu_sayisi','yas_grubu','longitude','latitude','tarih']
metro.head(3)

metro['longitude'] = metro['longitude'].str.replace('"', '')
metro['latitude'] = metro['latitude'].str.replace('"', '')
metro.head(3)

metro.isnull().sum()

empty_longitude = metro[metro['longitude'] == '']
empty_latitude = metro[metro['latitude'] == '']

metro.loc[metro['longitude'] == '', 'longitude'] = 0.0
metro.loc[metro['latitude'] == '', 'latitude'] = 0.0

metro['gecis_sayisi'] = metro['gecis_sayisi'].astype(int)
metro['yolcu_sayisi'] = metro['yolcu_sayisi'].astype(int)

metro['longitude'] = pd.to_numeric(metro['longitude'], errors='coerce')
metro['latitude'] = pd.to_numeric(metro['latitude'], errors='coerce')

metro.head(3)

metro.to_csv('../Datasets/PROCESSED/metro.csv', index=False)

<hr>

nufus_artis = pd.read_excel('../Datasets/EXTERNAL/TUIK-il ve ilcelere gore il_ilce merkezi belde_koy nufusu ve yillik nufus artis hizi.xls')
nufus_artis.head(3)

nufus_artis_hizi_istanbul = nufus_artis.iloc[np.where(nufus_artis.iloc[:,0] == 'İstanbul')[0][0]:503,:]

nufus_artis_hizi_istanbul.head(3)

nufus_artis_hizi_istanbul.columns = ["il_ilce", 'toplam_nufus', 'il_ilce_nufus', "belde_koy_nufus", "gec", "yillik_nufus_artis_hizi"]
nufus_artis_hizi_istanbul = nufus_artis_hizi_istanbul[['il_ilce', 'yillik_nufus_artis_hizi']]
nufus_artis_hizi_istanbul.head(3)

# turkish upper case
nufus_artis_hizi_istanbul['il_ilce'] = nufus_artis_hizi_istanbul['il_ilce'].str.upper()
nufus_artis_hizi_istanbul.head(3)

turkish_chars = {
    'ü': 'u',
    'Ü': 'U',
    'ğ': 'g',
    'Ğ': 'G',
    'ş': 's',
    'Ş': 'S',
    'ı': 'i',
    'İ': 'I',
    'ö': 'o',
    'Ö': 'O',
    'ç': 'c',
    'Ç': 'C'
}

from unidecode import unidecode

nufus_artis_hizi_istanbul['il_ilce'] = nufus_artis_hizi_istanbul['il_ilce'].str.replace('[^\w\s]', '', regex=True).apply(unidecode)

nufus_artis_hizi_istanbul.head(3)

nufus_artis_hizi_istanbul.to_csv('../Datasets/PROCESSED/nufus_artis_hizi_istanbul.csv', index=False)

<hr>

konut = pd.read_excel('../Datasets/EXTERNAL/TUIK-ilcelere gore konut satis sayilari.xls', skiprows=1)
konut.head(10)

np.where(konut.iloc[:,1] == 'İstanbul')[0][0]

np.where(konut.iloc[np.where(konut.iloc[:,1] == 'İstanbul')[0][0]:,1] != "İstanbul")[0][0]

konut = konut.iloc[228:228+39,[1,2,3]]
konut.columns = ['il', 'ilce', 'konut_satis_sayisi']

konut.head(3)

konut.to_csv('../Datasets/PROCESSED/konut_satis_sayilari.csv', index=False)

<hr>

park = pd.read_excel('../Datasets/EXTERNAL/park-ve-yeil-alan-koordinatlar.xlsx')
park.head()

park['latitude'] = ''
park['longitude'] = ''

for row in range(len(park)):
    park.loc[row, 'latitude'] = park.iloc[row,-3].split(',')[0]
    park.loc[row, 'longitude'] = park.iloc[row,-3].split(',')[1]

park.head(3)

park.loc[park['TÜR'] == 'METRO ÇIKIŞLARI', 'TÜR'] = 'METRO ÇIKIŞI'

park.columns = ['SIRA NO', 'TÜR', 'MAHAL ADI', 'İLÇE', 'KOORDİNAT',
       'latitude', 'longitude']

park['TÜR'].unique()

park.drop(['SIRA NO', 'KOORDİNAT'], axis=1, inplace=True)
park.head(3)

park.to_csv('../Datasets/PROCESSED/park.csv', index=False)
