==> Final-Solution/Notebooks/3-processed-datas.ipynb <==
import pandas as pd
import numpy as np

Get neighborhood from lat & long

my_user_agent = 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/112.0.0.0 Safari/537.36'

from geopy.geocoders import Nominatim

geolocator = Nominatim(user_agent=my_user_agent)

def get_district_neighborhood(latitude, longitude):
    
    location = geolocator.reverse(f"{latitude}, {longitude}")
    address = location.raw['address']

    # country
    country = address.get('country', '')
    
    # city
    city = address.get('province', '')
    
    
    # district
    district = address.get('town', '')
    
    # neighborhood
    neighborhood = address.get('suburb', '').replace(' Mahallesi', '').upper()
        
    # zipcode
    zipcode = address.get('postcode', '')

    
    df = pd.DataFrame({'country': [country],
                        'city': [city],
                            'district': [district],
                                'neighborhood': [neighborhood],
                                    'zipcode': [zipcode]})
    
    return district, neighborhood

<hr>

turkish_cases = {
    'İ': 'I',
    'ı': 'i',
    'Ğ': 'G',
    'ğ': 'g',
    'Ü': 'U',
    'ü': 'u',
    'Ş': 'S',
    'ş': 's',
    'Ö': 'O',
    'ö': 'o',
    'Ç': 'C',
    'ç': 'c',
}

database = pd.read_csv('../Datasets/PROCESSED/sosyal_yardim.csv')
database.columns = ['ilce', 'mahalle', 'yardim_alan_hane_sayisi']
database.head(3)

ses = pd.read_csv('../Datasets/PROCESSED/ses_skorlari.csv')
ses.columns = ['ilce', 'mahalle','ses_klasmanı', 'ses_skoru']
ses.head(3)

ses['ilce'] = ses['ilce'].str.upper()
ses['mahalle'] = ses['mahalle'].str.upper()

ses['ilce'] = ses['ilce'].apply(lambda x: ''.join([turkish_cases.get(i, i) for i in x]))
ses['mahalle'] = ses['mahalle'].apply(lambda x: ''.join([turkish_cases.get(i, i) for i in x]))

one_matched = []
multiple_match = []
non_matched = []
for i in ses['ilce'].unique().tolist():
    ct = 0
    for j in database['ilce'].unique():
        if i == j:
            ct += 1
    if ct > 1:
        multiple_match.append(i)
    elif ct == 1:
        one_matched.append(i)
    else:
        non_matched.append(i)

one_matched, multiple_match, non_matched

ses.loc[ses['ilce'] == 'EYUP', 'ilce'] = 'EYUPSULTAN'

database = pd.merge(database, ses, on=['ilce', 'mahalle'], how='left')
database.head(3)

park = pd.read_csv('../Datasets/PROCESSED/park.csv')
park.dropna(inplace=True)
park.head(3)

park['TÜR'].unique()

parklar = park[park['TÜR'] == 'PARK']
parklar.head(3)

parklar['İLÇE'].unique()

parklar = parklar[['İLÇE', 'TÜR']].groupby(['İLÇE', 'TÜR']).size().reset_index(name='count')
parklar.head(3)

metrolar = park[park['TÜR'] == 'METRO ÇIKIŞI']
metrolar.head(3)

metrolar['İLÇE'].unique()

metrolar = metrolar[['İLÇE', 'TÜR']].groupby(['İLÇE', 'TÜR']).size().reset_index(name='count')
metrolar.head(3)

park_metro_df = pd.concat([parklar, metrolar], axis=0)
park_metro_df.head(3)

park_metro_df['İLÇE'] = park_metro_df['İLÇE'].str.upper()
park_metro_df['İLÇE'] = park_metro_df['İLÇE'].apply(lambda x: ''.join([turkish_cases.get(i, i) for i in x]))
park_metro_df.head(3)

park_metro_df.columns = ['ilce', 'tur', 'tur_count']

database = pd.merge(database, park_metro_df, on=['ilce'], how='left')
database.head(3)

database.sort_values('tur_count', ascending=False).head(3)

nufus_maaliyet = pd.read_csv('../Datasets/PROCESSED/nufus_maaliyet.csv')
nufus_maaliyet.head(3)

database['mahalle'] = database['mahalle'].str.replace(' ', '')
database.head(10)

nufus_maaliyet['mahalle'] = nufus_maaliyet['mahalle'].str.replace(' ', '')
nufus_maaliyet.head(3)

database = pd.merge(database, nufus_maaliyet, on=['ilce', 'mahalle'], how='inner')

database.sort_values('maliyet', ascending=False).head(3)
