==> Final-Solution/Notebooks/6-subelerin-konumlandırılması.ipynb <==
import pandas as pd 
import numpy as np 

df = pd.read_csv("../Datasets/PROCESSED/database.csv")
df.head()

df = df[['İLÇE ADI','MAHALLE ADI', 'maliyet', 'agirlikli_puan']]
df.columns = ['ilce','mahalle','maaliyet','puan']
df.sort_values(by=['puan','maaliyet'], ascending=[False,True], inplace=True)
df.head(10)

import plotly.express as px

fig = px.box(df.head(20), x='puan', points='all')
fig.show()

<hr>

data = df.copy()

data.head()

data[data['puan']>data['puan'].quantile(0.85)]

data['karlılık'] = data['puan']*1000000 / data['maaliyet']

fig = px.scatter(data, x='maaliyet', y='puan', color='ilce', size='karlılık')
fig.show()

data = data[(data['puan']>0.4) & (data['maaliyet']>250000) & (data['maaliyet']<550000) & (data['karlılık']>1.0)]

data.sort_values(by=['puan', 'karlılık'], ascending=[0,0], inplace=True)

data.head(10)

choosen = pd.DataFrame()
budget = 3000000
for row in range(len(data)):
    budget -= data.iloc[row]['maaliyet']
    if budget <= 0:
        budget += data.iloc[row]['maaliyet']
        break
    else:
        choosen = pd.concat([choosen, data.iloc[[row]]], axis=0)

budget

choosen

fig = px.scatter(choosen, x='maaliyet', y='puan', color='ilce', size='karlılık')
fig.update_layout(title='Seçilen Mahalleler', height=600, width=900)
fig.show()

choosen.to_csv('../Datasets/PROCESSED/choosen.csv', index=False)
