==> Final-Solution/Notebooks/7-roi.ipynb <==
import pandas as pd 
import numpy as np

df = pd.read_csv('../Datasets/PROCESSED/choosen.csv')
df = df[['ilce','mahalle','maaliyet']]
df

df2 = pd.read_csv('../Datasets/PROCESSED/nufus_maaliyet.csv')

df2[df2['ilce']=='USKUDAR'].head()

müsteriler = [8311, 4033, 2562, 10522, 4785, 4860]

df['musteri'] = müsteriler
df

Brew Coffe House menüsü espresso bazlı içeceklerin ortalama fiyatı 50₺
